requirements.txt
anyio==4.9.0
certifi==2025.6.15
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.10
python-telegram-bot==22.2
sniffio==1.3.1

cryptography~=45.0.5
SQLAlchemy~=2.0.41